requirements.txt
requests
beautifulsoup4
python-dotenv
sendgrid
schedule
sqlite3